Final polish: Fabric system schemas, single local logging site, co-located resolvers
- FabricSqlDialect.is_system_schema extends (not replaces) the base
  information_schema exclusion: discovery now includes views, so Fabric
  was the one dialect where INFORMATION_SCHEMA objects would surface as
  discoverable datasets.
- The local discovery handler no longer catch-logs: failures propagate
  raw and the local wiring — the single logging site — maps them to
  LOG_ERRORS, so connection-open failures are no longer mislabeled as
  query failures and the wiring's generic arm is reachable again.
- resolve_scan_definition_name moves to handlers/dependencies.py next to
  the other resolvers (its tests move to test_cli_dependencies.py).
- discovery_payload reads the scan id through EnvConfigHelper, the
  canonical accessor.
- The DTO's type discriminator is Literal["sodaCoreInsertScanResults"].
