+++ soda-tests/tests/unit/test_cli_data_source_handlers.py
@@ -261,8 +261,8 @@ def test_discover_unexpected_post_query_exception_propagates(mock_discovery_run_
 
 
 # Local discovery handler tests. Unlike the Cloud sibling there is no scan
-# lifecycle and no failure reporting: failures are logged and map straight to
-# LOG_ERRORS.
+# lifecycle and no failure reporting: failures propagate raw and the CLI
+# wiring — the single logging site — maps them to LOG_ERRORS.
 
 
 @patch("soda_core.discovery.discovery_run.DiscoveryRun")
@@ -301,14 +301,16 @@ def test_discover_locally_passes_include_exclude_filters(mock_discovery_run_cls)
 
 
 @patch("soda_core.discovery.discovery_run.DiscoveryRun")
-def test_discover_locally_query_failure_exits_log_errors(mock_discovery_run_cls, caplog):
+def test_discover_locally_query_failure_propagates_raw(mock_discovery_run_cls, caplog):
     data_source_impl = _data_source_impl_fake()
     mock_discovery_run_cls.execute.side_effect = Exception("query failed")
 
-    exit_code = handle_discover_data_source_locally(data_source_impl)
+    with pytest.raises(Exception, match="query failed"):
+        handle_discover_data_source_locally(data_source_impl)
 
-    assert exit_code == ExitCode.LOG_ERRORS
-    assert any("query failed" in record.getMessage() for record in caplog.records)
+    # Nothing is logged at the raise site: the CLI wiring is the single logging
+    # site and maps the failure to LOG_ERRORS (see the wiring tests).
+    assert not any("query failed" in record.getMessage() for record in caplog.records)
     # The connection is always released, also on failure.
     data_source_impl.close_connection.assert_called_once()
